+++ src/mastodon_api/microsyntax/parser.rs
@@ -1,12 +1,22 @@
 use regex::Match;
 
-pub fn is_inside_code_block(match_: &Match, text: &str) -> bool {
+fn is_inside_tag(match_: &Match, tag: &str, text: &str) -> bool {
     // TODO: remove workaround.
     // Perform replacement only inside text nodes during markdown parsing
     let text_before = &text[0..match_.start()];
-    let code_open = text_before.matches("<code>").count();
-    let code_closed = text_before.matches("</code>").count();
-    code_open > code_closed
+    let tag_open = format!("<{tag}");
+    let tag_close = format!("</{tag}>");
+    let tag_open_count = text_before.matches(&tag_open).count();
+    let tag_close_count = text_before.matches(&tag_close).count();
+    tag_open_count > tag_close_count
+}
+
+pub fn is_inside_code_block(match_: &Match, text: &str) -> bool {
+    is_inside_tag(match_, "code", text)
+}
+
+pub fn is_inside_link(match_: &Match, text: &str) -> bool {
+    is_inside_tag(match_, "a", text)
 }
 
 #[cfg(test)]
@@ -21,6 +31,16 @@ mod tests {
         let mat = regexp.find(text).unwrap();
         assert_eq!(mat.start(), 9);
         let result = is_inside_code_block(&mat, text);
+        assert_eq!(result, true);
+    }
+
+    #[test]
+    fn test_is_inside_link() {
+        let text = r#"abc<a href="https://test">#tag</a>xyz<a>link</a>"#;
+        let regexp = Regex::new("#tag").unwrap();
+        let match_ = regexp.find(text).unwrap();
+        assert_eq!(match_.start(), 26);
+        let result = is_inside_link(&match_, text);
         assert_eq!(result, true);
     }
 }