+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Use `Like` activity to represent emoji reactions.
+
 ## [3.16.0] - 2025-02-03
 
 ### Added