+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Use `Like` activity to represent emoji reactions.
 
+### Fixed
+
+- Removed spurious warnings about vote content being not null.
+
 ## [3.16.0] - 2025-02-03
 
 ### Added