+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Use `Like` activity to represent emoji reactions.
+- Don't prune posts that contain polls with local votes.
 
 ### Fixed