+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Use `Like` activity to represent emoji reactions.
 - Don't prune posts that contain polls with local votes.
 - Reverse order of rules in `list-filter-rules` command output.
+- Don't allow hashtags inside links.
 
 ### Fixed