+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Parse custom emojis in poll options.
+
 ### Changed
 
 - Use `Like` activity to represent emoji reactions.